re-write of almost everything
