+++ src/lib.rs
@@ -1,93 +1,21 @@
 //! # Tacit
 //! Obvious, powerful logging focused on structure and simplicity.
 
-mod loggers;
+mod formatters;
+mod logger;
 mod outputs;
+mod properties;
 
+pub use crate::{formatters::*, logger::*, outputs::*, properties::*};
 use log::{Log, Metadata, Record};
-pub use loggers::*;
-pub use outputs::*;
-
-/// Describes all loggers provided by `tacit`
-pub trait Logger: Log {
-    /// Return the `LevelFilter` for the `Logger`
-    fn level_filter(&self) -> log::LevelFilter;
-
-    /// Set the `LevelFilter` for the `Logger`
-    fn set_level_filter(&mut self, level: log::LevelFilter);
-
-    /// Set the `LevelFilter` for the `Logger`, useful for chaining operations
-    fn with_level_filter(mut self, level: log::LevelFilter) -> Self
-    where
-        Self: Sized,
-    {
-        self.set_level_filter(level);
-        self
-    }
-
-    /// Set the `LevelFilter` for a particular module name.
-    fn set_module_level_filter(&mut self, module: String, level: log::LevelFilter);
-
-    /// Set the `LevelFilter` for a particular module name. Useful for chaining operations.
-    fn with_module_level_filter(mut self, module: String, level: log::LevelFilter) -> Self
-    where
-        Self: Sized,
-    {
-        self.set_module_level_filter(module, level);
-        self
-    }
-
-    /// Add a dynamic property to the logging output.
-    fn add_fn_prop(&mut self, name: String, prop: fn(&Record) -> StaticProperty);
-
-    /// Add a dynamic property to the logging output. Useful for chaining operations.
-    fn with_fn_prop(mut self, name: String, prop: fn(&Record) -> StaticProperty) -> Self
-    where
-        Self: Sized,
-    {
-        self.add_fn_prop(name, prop);
-        self
-    }
-
-    /// Add a static property to the logging output.
-    fn add_prop(&mut self, name: String, prop: StaticProperty);
-
-    /// Add a static property to the logging output. Useful for chaining operations.
-    fn with_prop(mut self, name: String, prop: StaticProperty) -> Self
-    where
-        Self: Sized,
-    {
-        self.add_prop(name, prop);
-        self
-    }
-
-    /// Only log from modules with an explicit module level filter, useful for quieting down
-    /// dependencies.
-    fn explicit_logging(&mut self);
-
-    /// Only log from modules with an explicit module level filter, useful for quieting down
-    /// dependencies.
-    fn with_explicit_logging(mut self) -> Self
-    where
-        Self: Sized,
-    {
-        self.explicit_logging();
-        self
-    }
-
-    /// Prepare the `Logger` for logging operations
-    fn finalize(self) -> Self
-    where
-        Self: Sized;
-}
 
 /// Main logger abstraction for Tacit. Combines one or more `Logger` implementations.
-pub struct TacitLogger {
-    loggers: Vec<Box<dyn Logger>>,
+pub struct TacitLogger<O: TacitOutput, F: TacitFormatter> {
+    loggers: Vec<Box<Logger<O, F>>>,
     max_level: log::LevelFilter,
 }
 
-impl Default for TacitLogger {
+impl<O: TacitOutput, F: TacitFormatter> Default for TacitLogger<O, F> {
     fn default() -> Self {
         Self {
             loggers: Vec::new(),
@@ -96,13 +24,13 @@ impl Default for TacitLogger {
     }
 }
 
-pub fn new() -> TacitLogger {
+pub fn new<O: TacitOutput, F: TacitFormatter>() -> TacitLogger<O, F> {
     TacitLogger::default()
 }
 
-impl TacitLogger {
+impl<O: 'static + TacitOutput, F: 'static + TacitFormatter> TacitLogger<O, F> {
     /// Add a logger to the pile
-    pub fn with_logger<L: 'static + Logger>(mut self, logger: L) -> Self {
+    pub fn with_logger(mut self, logger: Logger<O, F>) -> Self {
         self.max_level = std::cmp::max(self.max_level, logger.level_filter());
         self.loggers.push(Box::new(logger.finalize()));
         self
@@ -115,7 +43,7 @@ impl TacitLogger {
     }
 }
 
-impl Log for TacitLogger {
+impl<O: TacitOutput, F: TacitFormatter> Log for TacitLogger<O, F> {
     fn flush(&self) {}
 
     fn enabled(&self, metadata: &Metadata) -> bool {
@@ -125,64 +53,12 @@ impl Log for TacitLogger {
     fn log(&self, record: &Record) {
         if self.enabled(record.metadata()) {
             for logger in &self.loggers {
-                logger.log(record);
+                if logger.enabled(record.metadata()) {
+                    logger.log(record);
+                }
             }
         }
     }
-}
-
-/// Property to add to the log output
-pub enum StaticProperty {
-    String(String),
-    Number(i64),
-    Null,
-}
-
-impl From<String> for StaticProperty {
-    fn from(s: String) -> Self {
-        Self::String(s)
-    }
-}
-
-impl From<&str> for StaticProperty {
-    fn from(s: &str) -> Self {
-        Self::String(s.to_string())
-    }
-}
-
-impl From<Option<String>> for StaticProperty {
-    fn from(s: Option<String>) -> Self {
-        match s {
-            Some(s) => Self::String(s),
-            None => Self::Null,
-        }
-    }
-}
-
-impl From<Option<&str>> for StaticProperty {
-    fn from(s: Option<&str>) -> Self {
-        match s {
-            Some(s) => Self::String(s.to_string()),
-            None => Self::Null,
-        }
-    }
-}
-
-impl From<i64> for StaticProperty {
-    fn from(n: i64) -> Self {
-        Self::Number(n)
-    }
-}
-
-impl From<u32> for StaticProperty {
-    fn from(n: u32) -> Self {
-        Self::Number(n as i64)
-    }
-}
-
-pub enum Property {
-    Static(StaticProperty),
-    Function(Box<dyn Fn(&Record) -> StaticProperty + Send + Sync>),
 }
 
 #[cfg(test)]
